[REF] Use `torch.fx.Transformer` for jet overloading (#100)
* [REF] Replace manual node replacement with Transformer-based logic

This commit introduces the JetTransformer that is used to transform a given traced GraphModule
to a GraphModule where all on-placeholder-dependent-nodes are replaced with its jet counterpart.
Nodes that only depent on constant nodes, are (more or less) remain the same.

Note: Names of jet transformed nodes are changed; Names of constant-dependent nodes remain unchanged.

* fixup! [REF] Replace manual node replacement with Transformer-based logic