Fix issue with nested tuple arguments that occur when nesting graphs
The problem was that the lambda used to create the trivial node, used `*args` as arguments, which creates a lambda that takes a tuple as args. This was correctly parsed by the signature parser, but lead to a wrong nested tuple. The change allows only single output functions, but the code the lines above also expects only a single output tensor. So should be fine for now.